add middlewares and finish reviews

+++ routes/api.php
@@ -9,6 +9,9 @@ use App\Http\Controllers\orders\cartController;
 use App\Http\Controllers\orders\orderController;
 use App\Http\Controllers\reviews\blogConteroller;
 use App\Http\Middleware\checkAdmin;
+use App\Http\Middleware\CheckBelongsTo;
+use App\Http\Middleware\ownCart;
+use App\Http\Middleware\ownOrder;
 use Illuminate\Support\Facades\Route;
 
 use App\Http\Controllers\auth\AuthController;
@@ -16,7 +19,7 @@ use App\Http\Controllers\auth\AuthController;
 Route::group([
     'middleware' => 'api',
     'prefix' => 'auth'
-], function ($router) {
+], function () {
     Route::post('/register', [AuthController::class, 'register']);
     Route::post('/login', [AuthController::class, 'login']);
     Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:api');
@@ -26,60 +29,61 @@ Route::group([
 });
 
 Route::middleware(checkAdmin::class)->group(function () {
-    Route::controller(brandController::class)->group(function () {
-        Route::get('/brands/{lang}', 'index');
-        Route::get('/brand/{lang}/{id}', 'show');
-        Route::post('/brand/{lang}', 'store');
-        Route::post('/brand/{lang}/{id}', 'update');
-        Route::delete('/brand/{lang}/{id}', 'destroy');
-        Route::get('/brand/{lang}/{id}/products', 'products');
-    });
-
-    Route::controller(categoreyController::class)->group(function () {
-        Route::get('/categories/{lang}', 'index');
-        Route::get('/category/{lang}/{id}', 'show');
-        Route::post('/category/{lang}', 'store');
-        Route::post('/category/{lang}/{id}', 'update');
-        Route::delete('/category/{lang}/{id}', 'destroy');
-        Route::get('/category/{lang}/{id}/products', 'products');
-    });
-
-    Route::controller(productController::class)->group(function () {
-        Route::get('/products', 'index');
-        Route::get('/product/{id}', 'show');
-        Route::post('/product', 'store');
-        Route::post('/product/{id}', 'update');
-        Route::delete('/product/{id}','destroy');
-        Route::get('/product/{id}/colors', 'colors');
-        Route::get('/product/{id}/sizes', 'sizes');
-        Route::get('/product/{id}/colorSizes', 'colorSizes');
-    });
 
     Route::controller(colorSizesController::class)->group(function () {
-        Route::post('/addColor', 'addColor');
-        Route::post('/addSize', 'addSize');
-        Route::get('/colors', 'getAllColors');
-        Route::get('/sizes', 'getAllSizes');
+        Route::post('/addColor/{lang}', 'addColor');
+        Route::post('/addSize/{lang}', 'addSize');
+        Route::get('/colors/{lang}', 'getAllColors');
+        Route::get('/sizes/{lang}', 'getAllSizes');
     });
 
 });
 
+Route::controller(brandController::class)->group(function () {
+    Route::get('/brands/{lang}', 'index');
+    Route::get('/brand/{lang}/{id}', 'show');
+    Route::post('/brand/{lang}', 'store')->middleware(checkAdmin::class);
+    Route::post('/brand/{lang}/{id}', 'update')->middleware(checkAdmin::class);
+    Route::delete('/brand/{lang}/{id}', 'destroy')->middleware(checkAdmin::class);
+    Route::get('/brand/{lang}/{id}/products', 'products');
+});
+
+Route::controller(categoreyController::class)->group(function () {
+    Route::get('/categories/{lang}', 'index');
+    Route::get('/category/{lang}/{id}', 'show');
+    Route::post('/category/{lang}', 'store')->middleware(checkAdmin::class);
+    Route::post('/category/{lang}/{id}', 'update')->middleware(checkAdmin::class);
+    Route::delete('/category/{lang}/{id}', 'destroy')->middleware(checkAdmin::class);
+    Route::get('/category/{lang}/{id}/products', 'products');
+});
+
+
+Route::controller(productController::class)->group(function () {
+    Route::get('/products', 'index');
+    Route::get('/product/{id}', 'show');
+    Route::post('/product', 'store')->middleware(checkAdmin::class);
+    Route::post('/product/{id}', 'update')->middleware(checkAdmin::class);
+    Route::delete('/product/{id}','destroy')->middleware(checkAdmin::class);
+    Route::get('/product/{id}/colors', 'colors');
+    Route::get('/product/{id}/sizes', 'sizes');
+    Route::get('/product/{id}/colorSizes', 'colorSizes');
+});
 
 
 Route::controller(cartController::class)->group(function () {
     Route::post('/cart', 'addToCart');
-    Route::get('/cart', 'getCartItems');
-    Route::delete('/cart', 'deleteFromCart');
-    Route::delete('/cart/clear', 'clearCart');
+    Route::get('/cart', 'getCartItems')->middleware(ownCart::class);
+    Route::delete('/cart', 'deleteFromCart')->middleware(ownCart::class);
+    Route::delete('/cart/clear', 'clearCart')->middleware(ownCart::class);
 });
 
 Route::controller(orderController::class)->group(function () {
     Route::get('/orders', 'index')->middleware(checkAdmin::class);
     Route::get('/order/{id}', 'show');
     Route::post('/order', 'store');
-    Route::get('/user/orders', 'getUserOrders');
-    Route::post('/order/{id}/status', 'ChangeStatus');
-    Route::delete('/order/{id}', 'destroy');
+    Route::get('/user/orders', 'getUserOrders')->middleware(ownOrder::class);
+    Route::post('/order/{id}/status', 'ChangeStatus')->middleware(ownOrder::class);
+    Route::delete('/order/{id}', 'destroy')->middleware(ownOrder::class);
 });
 
 Route::controller(MailController::class)->group(function () {
@@ -91,7 +95,9 @@ Route::controller(MailController::class)->group(function () {
 Route::controller(blogConteroller::class)->group(function () {
     Route::get('/blogs', 'index');
     Route::get('/blog/{id}', 'show');
-    Route::post('/blog', 'store');
-    Route::post('/blog/{id}', 'update');
-    Route::delete('/blog/{id}', 'destroy');
+    Route::get('/user/blogs', 'userBlogs')->middleware(CheckBelongsTo::class);
+    Route::post('/product/{id}/blog', 'store');
+    Route::post('blog/{id}', 'update')->middleware(CheckBelongsTo::class);
+    Route::delete('/blog/{id}', 'destroy')->middleware(CheckBelongsTo::class);
+    Route::get('/product/{id}/blogs', 'productBlogs');
 });
